+++ tests/test_gce.py
@@ -1,0 +1,29 @@
+from unittest.mock import Mock
+
+import pytest
+
+from hart.exceptions import UserError
+from hart.providers import gce
+
+
+def test_get_selected_or_default_subnet_none():
+    with pytest.raises(UserError):
+        gce.get_selected_or_default_subnet([], 'irrelevant')
+
+
+def test_get_selected_or_default_subnet_multiple():
+    first = Mock()
+    first.name = 'first'
+    second = Mock()
+    second.name = 'second'
+    ret = gce.get_selected_or_default_subnet([first, second], 'first')
+    assert ret.name == 'first'
+
+
+def test_get_selected_or_default_subnet_none_matching():
+    first = Mock()
+    first.name = 'first'
+    second = Mock()
+    second.name = 'second'
+    with pytest.raises(UserError):
+        gce.get_selected_or_default_subnet([first, second], 'third')